# record-controlled-work-api/open-api-specification.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Record Controlled Work API
paths:
  /api/v1/applications:
    get:
      tags:
        - applications
      summary: Get all applications
      operationId: getApplications
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 25
        - name: officeId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applications"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '500':
          description: 'Internal server error'
    post:
      tags:
        - applications
      summary: Create an application
      operationId: createApplication
      requestBody:
        description: create an application
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequestBody'
      responses:
        '201':
          description: 'Created'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApplicationResponseBody"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'

  /api/v1/applications/{id}:
    get:
      tags:
        - applications
      summary: Get application by ID
      operationId: getApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error

  /api/v1/applications/{id}/means:
    put:
      tags:
        - applications
      summary: Update the means data for an application
      operationId: updateApplicationMeans
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeansDataRequestBody"
      responses:
        '204':
          description: 'Application means data updated successfully'
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '409':
          description: 'Conflict'
        '500':
          description: 'Internal server error'

  /api/v1/items:
    get:
      tags:
        - items
      summary: Get items
      operationId: getItems
      responses:
        '200':
          description: 'Success'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Items"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
    post:
      tags:
        - items
      summary: Create an item
      operationId: createItem
      requestBody:
        description: create an item
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequestBody'
      responses:
        '201':
          description: 'Created'
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'

  /api/v1/items/{id}:
    get:
      summary: Get item by id
      operationId: getItemById
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the item to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
    put:
      tags:
        - items
      summary: Update an item
      operationId: updateItem
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the item to be updated
          schema:
            type: integer
            format: int64
      requestBody:
        description: update an existing item
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequestBody'
      responses:
        '204':
          description: No content
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'
    delete:
      summary: Delete item by id
      operationId: deleteItem
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the item to be deleted
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No content
        '400':
          description: 'Bad request'
        '401':
          description: 'Unauthorized'
        '403':
          description: 'Forbidden'
        '404':
          description: 'Not found'
        '500':
          description: 'Internal server error'

components:
  schemas:
    ApplicationOverview:
      required:
        - id
        - name
        - applicationRefNumber
        - modifiedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        applicationRefNumber:
          type: string
        modifiedAt:
          type: string
          format: date-time

    CreateApplicationRequestBody:
      required:
        - ecfFlag
        - legalAidBefore
        - clientDetails
      properties:
        ecfFlag:
          type: boolean
        legalAidBefore:
          type: string
        legalAidLast6Months:
          type: boolean
        reasonForReapplication:
          type: string
        clientDetails:
          $ref: "#/components/schemas/ClientDetails"

    CreateApplicationResponseBody:
      properties:
        id:
          type: string
          format: uuid
        ecfFlag:
          type: boolean
        legalAidBefore:
          type: string
        legalAidLast6Months:
          type: boolean
        reasonForReapplication:
          type: string
        clientDetails:
          $ref: "#/components/schemas/ClientDetails"

    Applications:
      type: array
      items:
        $ref: "#/components/schemas/ApplicationOverview"

    Application:
      type: object
      required:
        - id
        - individualLegalAidNumber
        - providerFirmCode
        - providerOfficeCode
        - createdAt
        - createdBy
        - modifiedAt
        - modifiedBy
        - clientDetails
        - applicationState
        - ecfFlag
        - applicationType
      properties:
        id:
          type: string
          format: uuid
        individualLegalAidNumber:
          type: string
          format: uuid
        providerFirmCode:
          type: string
        providerOfficeCode:
          type: string
        meansAssessmentId:
          type: string
          format: uuid
        clientDetails:
          $ref: "#/components/schemas/ClientDetails"
        applicationStatus:
          $ref: "#/components/schemas/ApplicationStatus"
        declaration:
          $ref: "#/components/schemas/Declaration"
        evidence:
          $ref: "#/components/schemas/Evidence"
        reasonForReapplication:
          type: string
        meansAssessmentRequired:
          type: boolean
        typeOfNonMeans:
          type: boolean
        ecfFlag:
          type: boolean
        contribution:
          type: string
        applicationType:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        modifiedAt:
          type: string
          format: date-time
        modifiedBy:
          type: string

    UpdateMeansDataRequestBody:
      type: object
      required:
        - data
        - result
      properties:
        data:
          type: object
        result:
          type: object

    Declaration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientDeclarationStatus:
          $ref: "#/components/schemas/ClientDeclarationStatus"
        declarationConfirmation:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        modifiedAt:
          type: string
          format: date-time
        modifiedBy:
          type: string

    Evidence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        evidenceStatus:
          $ref: "#/components/schemas/EvidenceStatus"
        payeIncomeEvidence:
          type: boolean
        otherIncomeEvidence:
          type: boolean
        housingCostsEvidence:
          type: boolean
        capitalEvidence:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        modifiedAt:
          type: string
          format: date-time
        modifiedBy:
          type: string

    ClientDetails:
      required:
        - firstName
        - lastName
        - dateOfBirth
        - hasFixedAddress
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        niNumber:
          type: string
        hasFixedAddress:
          type: boolean
        address:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time

    Address:
      required:
        - addressLine1
        - country
      type: object
      properties:
        id:
          type: string
          format: uuid
        addressLine1:
          type: string
        addressLine2:
          type: string
        addressLine3:
          type: string
        addressLine4:
          type: string
        townOrCity:
          type: string
        postCode:
          type: string
        county:
          type: string
        country:
          type: string
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time

    ClientDeclarationStatus:
      type: string
      enum:
        - DRAFT
    EvidenceStatus:
      type: string
      enum:
        - DRAFT
    ApplicationStatus:
      type: string
      enum:
        - DRAFT
        - COMPLETE


    Item:
      required:
        - id
        - name
        - description
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    ItemRequestBody:
      required:
        - name
        - description
      properties:
        name:
          type: string
        description:
          type: string

    Items:
      type: array
      items:
        $ref: "#/components/schemas/Item"